refactor(parts): lifecycle manager to its own file
